Fixed generation bug. Updated game to allow removal of structures. Added more routers and cables

+++ src/main/java/game/objects/infrastructure/SmallRouter.java
@@ -5,6 +5,6 @@ import com.raylib.Jaylib;
 public class SmallRouter {
     //The time taken between packet releases.
     //The smaller the number, the more packers per second can be handles
-    public static int packetHandleDelay = 1;
+    public static float packetHandleDelay = 1f;
     public static int maxCacheSize = 5;
 }
